% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold.R
\name{fold}
\alias{fold}
\title{Supporting function - fold()}
\usage{
fold(train, test, model1, model2, vary, varx = NA)
}
\arguments{
\item{train}{Data frame. Training dataset.}

\item{test}{Data frame. Testing dataset.}

\item{model1}{Character. The first model (lavaan syntax).}

\item{model2}{Character. The second model (lavaan syntax).}

\item{vary}{Character vector. Names of observed variables.}

\item{varx}{Character. Optional. Names of covariates (default = `NA`).}
}
\value{
A named vector of IMV values for each variable and a summary value ("all").
}
\description{
Performs cross-validation to compute IMV for two models.
}
\examples{
# Example usage is covered by higher-level functions imvsem and imv_baseline.
}
