% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imvsem.R
\name{imvsem}
\alias{imvsem}
\title{Calculate IMV Using Two Models}
\usage{
imvsem(model1, model2 = NULL, vary, data, seed = 1234, nfold = 5)
}
\arguments{
\item{model1}{Character. The first model (lavaan syntax).}

\item{model2}{Character. The second model (lavaan syntax).
when model2 = NULL, the comparison will be conducted between model1 and a baaseline model}

\item{vary}{Character vector. Names of observed variables.}

\item{data}{Data frame. The dataset used for training/testing.}

\item{seed}{Integer. Seed for random sampling (default = `1234`).}

\item{nfold}{Integer. Number of folds for cross-validation (default = `5`,
when nfold = 1, the IMV will be based on insample prediction instead of out of sample prediction).}
}
\value{
A matrix of baseline IMV values for each variable and scale-level imv values ("all").
}
\description{
Computes the Information Measure for Validation (IMV) by comparing two models
across specified variables using cross-validation or in-sample prediction.
}
\examples{
# Load necessary libraries
library(lavaan)
library(stats)
library(imv)

# Load the dataset from the inst/extdata folder
data_path <- system.file("extdata", "BPI.dat", package = "imv4sem")
data <- read.table(data_path, header = TRUE)

# Define the variables to be analyzed
vary <- colnames(data)

# Define the first model (6 factors)
model_6f <- '
AnxDep =~ NA*ad1 + ad2 + ad3 + ad4 + ad5
Headstr =~ NA*hs1 + hs2 + hs3 + hs4 + hs5
Antisoc =~ NA*as1 + as2 + as3 + as4 + as5 + as6
Hyperac =~ NA*hy1 + hy2 + hy3 + hy4 + hy5
PeerProb =~ NA*pp1 + pp2 + pp3
Depend =~ NA*de1 + de2 + de3 + de4
AnxDep ~~ 1*AnxDep
Headstr ~~ 1*Headstr
Antisoc ~~ 1*Antisoc
Hyperac ~~ 1*Hyperac
PeerProb ~~ 1*PeerProb
Depend ~~ 1*Depend
'

# Define the second model (5 factors)
model_5f <- '
AnxDep =~ NA*ad1 + ad2 + ad3 + ad4 + ad5
Headstr =~ NA*hs1 + hs2 + hs3 + hs4 + hs5
Antisoc =~ NA*as1 + as2 + as3 + as4 + as5 + as6
Hyperac =~ NA*hy1 + hy2 + hy3 + hy4 + hy5
PeerProb =~ NA*pp1 + pp2 + pp3 + de1 + de2 + de3 + de4
AnxDep ~~ 1*AnxDep
Headstr ~~ 1*Headstr
Antisoc ~~ 1*Antisoc
Hyperac ~~ 1*Hyperac
PeerProb ~~ 1*PeerProb
'

# Calculate IMV between 5-factor model and 6-factor model using out-of-sample prediction
imv56f <- imvsem(model1 = model_5f, model2 = model_6f, vary = vary, data = data, nfold = 5)
# Calculate IMV between 5-factor model and 6-factor model using in-sample prediction
imv56f_in <- imvsem(model1 = model_5f, model2 = model_6f, vary = vary, data = data, nfold = 1)
# Calculate IMV between 6-factor model and baseline model using out-of-sample prediction
imv6f_base <- imvsem(model1 = model_6f, vary = vary, data = data, nfold = 5)

}
