% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot4imv.R
\name{plot4imv}
\alias{plot4imv}
\title{Plot Item-Level IMV with Thresholds and Scale-Level Note}
\usage{
plot4imv(imvres)
}
\arguments{
\item{imvres}{Matrix or vector. Output from the `imvsem` or `imv_baseline` function.}
}
\value{
A ggplot object showing the item-level IMV bar chart with thresholds.
}
\description{
This function calculates the average IMV across all folds and creates a bar chart
for the item-level IMVs. The scale-level IMV is included as a note in the plot.
}
\examples{
# In-sample prediction IMV values
imvres <- as.matrix(t(c(ad1 = 0.02, ad2 = 0.03, ad3 = 0.05, ad4 = 0.01, all = 0.03)))
plot(imvres)

}
