% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.imv.R
\name{print.summary_imvsem}
\alias{print.summary_imvsem}
\title{Print Summary for IMV}
\usage{
\method{print}{summary_imvsem}(x, ...)
}
\arguments{
\item{x}{An object of class `summary_imvsem` obtained from summary.imv().}

\item{...}{Additional arguments (not used).}
}
\value{
printed results for IMV values
}
\description{
Custom print method for objects of class `summary_imvsem`.
}
