% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.imv.R
\name{summary_imvsem}
\alias{summary_imvsem}
\title{Summarize IMV Results}
\usage{
summary_imvsem(object, ...)
}
\arguments{
\item{object}{Matrix or data frame. The output from the `imvsem` or `imv_baseline` function.}

\item{...}{Additional arguments (not used).}
}
\value{
A list of class `summary_imvsem` containing the following components:
\item{scale_imv}{Scale-level IMV (average of the `all` column).}
\item{item_level_imv}{A data frame with item-level IMVs and their names.}
\item{n_items}{The number of item-level IMVs.}
}
\description{
This function provides a summary of IMV values for a given object.
}
\examples{
# Example IMV result
imvres <- as.matrix(t(c(ad1 = 0.02, ad2 = 0.03, ad3 = 0.05, ad4 = 0.01, all = 0.03)))
summary_imvsem <- summary(imvres)
print(summary_imvsem)

}
